Integrated astrobee common into astrobee (#43)
The files in the astrobee common repo have been intergrated into the
astrobee repo. The astrobee idls can now be found in
communications/dds_msgs/idl. The plan schema can be found in
astrobee/commands and the command dictionary can be found in
doc/general_documentation. Also, documentation on how to add a ground
command to the astrobee system has been added to the
doc/general_documentation folder.