Dock Position Check Timeout (#94)
When undocking, the robot switches to ml, and then checks localization to make sure it matches a berth position. If the robot takes some time after the switch to reacquire position this check fails and switches to no localization. This is a fix for this situation, where instead of failing right away, it allows some time to reacquire position.