Localization manager work (#137)
* Changed registration timeout for ml and ar.

With the new localization system, the registration messages aren't
published as frequently. The timeouts for ml and ar have been changed to
reflect this.

* Added localization manager state transitions.

The localization manager was only able to execute a reset ekf command
when the pipeline was localization, not when it was unstable. This has
been fixed. The localization manager was also changed to allow an
initialize bias command when Astrobee is not localizing.

* Added better debugging output statements.