Added gs topics to the default recording. (#261)
The dds ros bridge died during an activity and it looks like it may have
been due to a mismatch in the guest science state and config messages.
Since these topics weren't recorded, it is hard to piece together what
happened. Thus these topics were added to the default topics that get
recorded all the time.