Astrobee comms work (#504)
* Changed astrobee comms trigger service to enable.

* Changed response for start astrobee intercomms srv

The start astrobee intercommunication service now returns an error
message with what went wrong if the service fails. The astrobee to
astrobee bridge no longer kills itself if its config is wrong. This
allows an operator to fix the config and resend the enable astrobee
intercommunication command during run time.

* Added the enable astrobee intercomms command.

* Fixed lint error.

* Fixed major typo in apk print version script. (#499)