Added output to system monitor.
When the system monitor detects that it didn't receive an initial
heartbeat from a node in the system, it assumes the node didn't startup
and it tries to start it manually. This works great on the robot but in
simulation, it generates cryptic errors when trying to start hardware
nodes. The added output should help the user figure out what is going
on.